+++ CHANGELOG.md
@@ -11,16 +11,23 @@ This project uses a [CalVer](https://calver.org/) versioning scheme with monthly
 - Code implementation for communicating with Schunk grippers has been changed to `SchunkGripperProcess`, `SchunkEGK40_USB` has been relegated to the schunk branch of this repo.
 - Default manipulator specs in `URrtde` class were wrong, they are now multiplied by π -> **caution**: your robot may suddenly move 3x faster if you didn't set joint speed yourself.
 - Old airo-teleop package was removed from the monorepo, in favor of a new [airo-teleop](https://github.com/airo-ugent/airo-teleop/), which is now considered a sister repository.
+- Changed the usage interface of the parameter constants in `zed.py`. These constants are now accessible through their respective dataclass (e.g., NEURAL_DEPTH_MODE -> InitParams.NEURAL_DEPTH_MODE).
 
 ### Added
 - Added `gc_disabled()` context manager to temporarily disable garbage collection for performance-critical sections.
 - Added `get_model()` method to `URrtde` to automatically detect the UR model (UR3, UR3e, UR5e)
+- Added `_retrieve_camera_pose()` method to the `Zed` class in `zed.py` to retrieve the pose matrix of the camera.
+- Added `_request_spatial_map_update` and `_retrieve_spatial_map` methods to the `Zed` class in `zed.py` to respectively request and retrieve a spatial map update using VSLAM.
+- Added dataclasses for the different kinds of Zed parameters (init parameters, positional tracking and spatial mapping) and a `ZedSpatialMap` dataclass.
+- Added extra test code in the `__main__` block of `zed.py` to test the new methods.
+- Added a new file `multiprocess_zed_camera.py`, implementing ZED-specific publisher and receiver classes that extend the StereoRGBD publisher and receiver with tracking and spatial mapping data exchange over shared memory.
 
 ### Changed
 - Improved `execute_trajectory` reliability: we now temporarily disable the garbage collector in the hot loop to reduce latency using `gc_disabled()`.
 - Deprecated `AsyncExecutor`. It is a very thin wrapper around [`ThreadPoolExecutor(max_workers=1)`](https://docs.python.org/3/library/concurrent.futures.html#concurrent.futures.Executor), which should be used instead.
 - `get_model()` (see Added) is now used in the URrtde constructor to automatically detect the UR model and set the manipulator specs appropriately.
 - Dropped support for Python 3.9 and added support for 3.12. This only has an impact on our CI workflow and should not impact you directly, but it does mean that airo-mono could stop working without any warning on Python 3.9 in the future.
+- Changed incorrect usages of the updated parameter constants in `zed.py` throughout the `airo-camera-toolkit` (see Breaking changes).
 
 ### Fixed